requirement.txt
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.3.4
attrs==20.3.0
autobahn==21.2.1
Automat==20.2.0
autopep8==1.5.5
beautifulsoup4==4.9.1
botocore==1.20.44
cachetools==4.1.1
certifi==2020.11.8
cffi==1.14.4
channels==3.0.3
chardet==3.0.4
click==7.1.2
colorama==0.4.4
constantly==15.1.0
cryptography==3.0
daphne==3.0.1
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20200713
distlib==0.3.1
dj-database-url==0.5.0
Django==2.0
django-admin-rangefilter==0.6.3
django-allauth==0.44.0
django-better-admin-arrayfield==1.3.0
django-bootstrap-datepicker-plus==3.0.5
django-ckeditor==6.0.0
django-cleanup==5.1.0
django-cors-headers==3.7.0
django-crispy-forms==1.9.2
django-filter==2.3.0
django-gst-field==0.0.3
django-heroku==0.3.1
django-import-export==2.4.0
django-js-asset==1.2.2
django-phone-field==1.8.1
django-phonenumber-field==5.0.0
django-qrcode==0.3
django-queryset-csv==1.1.0
django-rest-framework==0.1.0
django-rest-knox==4.1.0
django-session-timeout==0.1.0
django-simple-history==2.11.0
django-stronghold==0.4.0
django-taggit==1.3.0
django-templated-mail==1.1.1
django-utils-six==2.0
django-widget-tweaks==1.4.8
djangorestframework==3.7.7
djangorestframework-api-key==2.0.0
djangorestframework-jsonapi==4.0.0
djangorestframework-jwt==1.11.0
djoser==2.0.5
dynaconf==3.1.2
env2==0.0.0
et-xmlfile==1.0.1
filelock==3.0.12
geographiclib==1.50
geopy==2.0.0
google==3.0.0
google-auth==1.20.1
google-oauth==1.0.1
gunicorn==20.0.4
html5lib==1.1
hvac==0.10.5
hyperlink==21.0.0
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
importlib-metadata==2.0.0
incremental==21.3.0
inflection==0.5.1
jdcal==1.4.1
Jinja2==2.11.2
jmespath==0.10.0
knox==0.1.0
loguru==0.5.3
Markdown==3.3.1
MarkupPy==1.14
MarkupSafe==1.1.1
moviepy==1.0.3
mysqlclient @ file:///E:/mysqlclient-1.4.6-cp37-cp37m-win32.whl
numpy==1.19.3
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.5
pandas==1.2.3
phonenumbers==8.12.10
Pillow==8.0.1
proglog==0.1.9
psycopg2==2.8.6
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pycryptodome==3.9.9
PyJWT==1.7.1
pymongo==3.11.0
PyMySQL==1.0.2
pyOpenSSL==19.1.0
PyPDF2==1.26.0
PyQRCode==1.2.1
pytesseract==0.3.7
python-dateutil==2.8.1
python-decouple==3.4
python3-openid==3.2.0
pytz==2020.4
PyYAML==5.3.1
qrcode==6.1
razorpay==1.2.0
reportlab==3.5.55
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
s3transfer==0.3.4
service-identity==18.1.0
shortuuid==1.0.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
stripe==2.55.1
tablib==2.0.0
toml==0.10.2
tqdm==4.48.2
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
typescript==0.0.12
typing-extensions==3.10.0.0
unicodecsv==0.14.1
urllib3==1.26.2
validate-email==1.3
virtualenv==20.1.0
virtualenvwrapper-win==1.2.6
webencodings==0.5.1
whitenoise==5.2.0
win32-setctime==1.0.3
xhtml2pdf==0.2.4
xlrd==1.2.0
xlwt==1.3.0
zipp==3.3.1
zope.interface==5.2.0
